In KafkaW::Msg use librdkafkas types for message properties

+++ src/KafkaW/Msg.cxx
@@ -8,7 +8,7 @@ Msg::~Msg() { rd_kafka_message_destroy((rd_kafka_message_t *)MsgPtr); }
 
 uchar *Msg::data() { return (uchar *)((rd_kafka_message_t *)MsgPtr)->payload; }
 
-uint32_t Msg::size() { return ((rd_kafka_message_t *)MsgPtr)->len; }
+size_t Msg::size() { return ((rd_kafka_message_t *)MsgPtr)->len; }
 
 char const *Msg::topicName() {
   return rd_kafka_topic_name(((rd_kafka_message_t *)MsgPtr)->rkt);
@@ -16,5 +16,5 @@ char const *Msg::topicName() {
 
 int32_t Msg::partition() { return ((rd_kafka_message_t *)MsgPtr)->partition; }
 
-int32_t Msg::offset() { return ((rd_kafka_message_t *)MsgPtr)->offset; }
+int64_t Msg::offset() { return ((rd_kafka_message_t *)MsgPtr)->offset; }
 }
